pyproject.toml
[tool.poetry]
name = "практическая работа 1"
version = "0.1.0"
description = ""
authors = ["gigafiga21 <gigafiga21@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
virtualenv = "^20.8.0"
requests = "^2.26.0"
beautifulsoup4 = "^4.10.0"
Scrapy = "^2.5.0"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
